file - organized components with new folders (landingPage, views, modals, customSelect)
