+++ components/forms/NPCForm.tsx
@@ -8,8 +8,8 @@ import { npcSchema } from '@/database/drizzle/validation';
 import { CheckboxGroup, Checkbox, Textarea, Button } from '@nextui-org/react';
 import { SubmitButton } from '@/components/buttons/SubmitButton';
 import ReactSelect from 'react-select';
-import { VoiceOption } from './dropdown/VoiceOption';
-import { VoiceSingleValue } from './dropdown/VoiceSingleValue';
+import { VoiceOption } from './customSelect/VoiceOption';
+import { VoiceSingleValue } from './customSelect/VoiceSingleValue';
 import { PlusIcon } from '../icons';
 import { ErrorMessage } from '@hookform/error-message';
 import { ErrorToast } from '@/components/ErrorToast';
