+++ components/views/CampaignDetail.tsx
